Some very minor refactoring.
