+++ utility/__init__.py
@@ -1,1 +1,5 @@
-# The existance of this file makes this subfolder a "package"
+# The existence of this file makes this subfolder a "package"
+"""
+This package contains general-purpose code that is not necessarily specific to this application.
+Don't be surprised if we move some/all of this code to an external library at some point.
+"""
